refactor: break up task-add-dialog into subcomponents (#23)
Extract TaskSubtasksEditor, TaskTagsInput, TaskDescriptionInput,
and TaskMetadataRow into src/features/tasks/components/task-form/.
Reduces task-add-dialog from 596 to ~280 lines.
