+++ src/features/tasks/components/task-form/index.ts
@@ -1,0 +1,4 @@
+export { TaskSubtasksEditor } from "./task-subtasks-editor";
+export { TaskTagsInput } from "./task-tags-input";
+export { TaskDescriptionInput } from "./task-description-input";
+export { TaskMetadataRow } from "./task-metadata-row";
